Python Programs/Video Processing in Python/requirements.txt
moviepy
pillow
ffmpeg
ImageMagick
